Added Text Scripts and Transition scripts.
Can explain more in class today, Had to do each indivdually for the opening dialouge, could not get typewriter effect for the texts howerver. all text can and wil be controlled by the game manager, disabling and turning on scripts to allow the text to flow. currently only the Mei Yin will work after starting the game adding the correct scripts and variables and pressing K followed by space to advance the text. I choose spacebar but it can be easily switched