Final Keypad and texts additions, (i have not added any of my changes to the main scenes.)
Final Keypad scene is new and is where the final keypad i have done up will work. I explained the working on that a little bit on the discord, but if needed lmk ill get you the description. With the text, I have them labeled of which text compnent goes to which NPC. I know it's confusing, so i am free basically at anytime to implement that exactly perfectly. It uses a boolean to see if you have talked to the person, If it does register as you talked it says you already talked to them. If not and you are in the collider, you initiate the opening conversation with the NPC. And by leaving the collider you are able to disable the script but i believe keep the knowledge of being talked to