Finished implementing
Everything should be working now and all the features are implemented. Can find and do everything. E is for interactions, Space is for talking to people, N to bring up notebook, escape used ocassionally. Sometimes there is a glitch that stops character from going through doors (not sure why), but just restarting the build fixes it